code/requirements.txt
# Requirements for Prompt Engineering Advanced Guide Code Examples
# ================================================================

# Core dependencies for API integration
groq>=0.4.0
python-dotenv>=1.0.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Text processing and similarity
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Testing and validation
pytest>=7.0.0
pytest-cov>=4.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For advanced examples
openai>=1.0.0  # Alternative API client
anthropic>=0.3.0  # Another alternative
